Refactor kode form verifikasi
Memisahkan form edit dan view menjadi komponen tersendiri
